Attempt to port to 1.7.10
Of course this port is not working yet. My plan is to get it to compile first, then fix this the occuring issues. I have disabled the mixin lib, I am not sure if this is gonna be temporary(if I find a mixin lib for 1.7.10) or permanent(I will replace it's implementations). Any help is appreciated, and preferably announced. If you have questions, ask.